requirements.txt
Flask==2.3.3
Pillow==10.0.0
reportlab==4.0.4
python-dotenv==1.0.0
flask-cors==4.0.0 